Remove Deleter from SharedPtr's & WeakPtr's template parameters
As before, type of the deleter is either the default deleter type
or a user-supplied type.

Changes:
* Deleter is now stored indirectly (in a control block)
* Reset operation can now give a SharedPtr a different deleter

Deleter remains a part of the UniquePtr's template instantiation type
because we want UniquePtr to be a zero-overhead abstraction.

This makes std::unique_ptr a zero-overhead abstraction, which means that
all calls to the deleter can be inlined.